% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_input}
\alias{check_input}
\title{Generic input checker}
\usage{
check_input(input, substitution)
}
\arguments{
\item{input}{string to check}

\item{substitution}{string for what's been looked for}
}
\value{
boolian indicating whether the checked string contains a match for the substitution string
}
\description{
Utility function to handle namespaces. Used by \code{get_item} and \code{get_property}
}
