% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createrows}
\alias{createrows}
\title{"CREATE" rows}
\usage{
createrows(items, vector)
}
\arguments{
\item{items}{a vector, data frame or tibble of items (which may or may not contain the keyword "CREATE")}

\item{vector}{a vector of properties or values which may be expanded based on the items vector}
}
\value{
if the vector is NULL, return NULL. Otherwise, if the "CREATE" keyword appears in the
items vector, insert blank strings at those positions in the vector.
}
\description{
Add in empty lines for QuickStatements CREATE rows that mint new QIDs.
This is a slightly messy quirk of the QuickStatements format that mints new QIDs via a line
containing only "CREATE", so this function is a way to approximate that bevaviour in a tibble
}
