% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_para}
\alias{extract_para}
\title{Extract a paragraph of text}
\usage{
extract_para(text, para = 1, templ = NULL)
}
\arguments{
\item{text}{the input text as a string}

\item{para}{number indicating which paragraph(s) to return (default=1)}

\item{templ}{an optional string specifying a mediawikitemplate within
which to restrict the search restrict search}
}
\value{
the nth paragraph of the input text.
}
\description{
Return the nth paragraph of a section of text
Useful for extracting information from wikipedia or other wikimarkup text
}
