% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{get_geo_entity}
\alias{get_geo_entity}
\title{Retrieve geographic information from Wikidata}
\usage{
get_geo_entity(entity, language = "en", radius = NULL, limit = 100, ...)
}
\arguments{
\item{entity}{a Wikidata item (\code{Q...}) or series of items, to check
for associated geo-tagged items.}

\item{language}{the two-letter language code to use for the name
of the item. "en" by default, because we're imperialist
anglocentric westerners.}

\item{radius}{optionally, a radius (in kilometers) around \code{entity}
to restrict the search to.}

\item{limit}{the maximum number of results to return.}

\item{...}{
  Arguments passed on to \code{\link[httr:GET]{httr::GET}}
  \describe{
    \item{\code{url}}{the url of the page to retrieve}
    \item{\code{config}}{Additional configuration settings such as http
authentication (\code{\link[httr:authenticate]{authenticate()}}), additional headers
(\code{\link[httr:add_headers]{add_headers()}}), cookies (\code{\link[httr:set_cookies]{set_cookies()}}) etc.
See \code{\link[httr:config]{config()}} for full details and list of helpers.}
    \item{\code{handle}}{The handle to use with this request. If not
supplied, will be retrieved and reused from the \code{\link[httr:handle_pool]{handle_pool()}}
based on the scheme, hostname and port of the url. By default \pkg{httr}
requests to the same scheme/host/port combo. This substantially reduces
connection time, and ensures that cookies are maintained over multiple
requests to the same host. See \code{\link[httr:handle_pool]{handle_pool()}} for more
details.}
  }}
}
\value{
a data.frame of 5 columns:
\describe{
\item{item}{ the Wikidata identifier of each object associated with
\code{entity}.}
\item{name}{ the name of the item, if available, in the requested language. If it
is not available, \code{NA} will be returned instead.}
\item{latitude}{ the latitude of \code{item}}
\item{longitude}{ the longitude of \code{item}}
\item{entity}{ the entity the item is associated with (necessary for multi-entity
queries).}
}
}
\description{
\code{get_geo_entity} retrieves the item ID, latitude
and longitude of any object with geographic data associated with \emph{another}
object with geographic data (example: all the locations around/near/associated with
a city).
}
\examples{
# All entities
\donttest{
sf_locations <- get_geo_entity("Q62")

# Entities with French, rather than English, names
sf_locations <- get_geo_entity("Q62", language = "fr")

# Entities within 1km
sf_close_locations <- get_geo_entity("Q62", radius = 1)

# Multiple entities
multi_entity <- get_geo_entity(entity = c("Q62", "Q64"))
}

}
\seealso{
\code{\link{get_geo_box}} for using a bounding box
rather than an unrestricted search or simple radius.
}
