% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{get_item}
\alias{get_item}
\alias{get_property}
\title{Retrieve specific Wikidata items or properties}
\usage{
get_item(id, ...)

get_property(id, ...)
}
\arguments{
\item{id}{the ID number(s) of the item or property you're looking for. This can be in
various formats; either a numeric value ("200"), the full name ("Q200") or
even with an included namespace ("Property:P10") - the function will format
it appropriately. This function is vectorized and will happily accept
multiple IDs.}

\item{...}{
  Arguments passed on to \code{\link[httr:GET]{httr::GET}}
  \describe{
    \item{\code{url}}{the url of the page to retrieve}
    \item{\code{config}}{Additional configuration settings such as http
authentication (\code{\link[httr:authenticate]{authenticate()}}), additional headers
(\code{\link[httr:add_headers]{add_headers()}}), cookies (\code{\link[httr:set_cookies]{set_cookies()}}) etc.
See \code{\link[httr:config]{config()}} for full details and list of helpers.}
    \item{\code{handle}}{The handle to use with this request. If not
supplied, will be retrieved and reused from the \code{\link[httr:handle_pool]{handle_pool()}}
based on the scheme, hostname and port of the url. By default \pkg{httr}
requests to the same scheme/host/port combo. This substantially reduces
connection time, and ensures that cookies are maintained over multiple
requests to the same host. See \code{\link[httr:handle_pool]{handle_pool()}} for more
details.}
  }}
}
\description{
\code{get_item} and \code{get_property} allow you to retrieve the data associated
with individual Wikidata items and properties, respectively. As with
other \code{WikidataR} code, custom print methods are available; use \code{\link{str}}
to manipulate and see the underlying structure of the data.
}
\examples{

#Retrieve a specific item
adams_metadata <- get_item("42")

#Retrieve a specific property
object_is_child <- get_property("P40")

}
\seealso{
\code{\link{get_random}} for selecting a random item or property,
or \code{\link{find_item}} for using search functionality to pull out
item or property IDs where the descriptions or aliases match a particular
search term.
}
