% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_names_from_properties}
\alias{get_names_from_properties}
\title{Get names of properties}
\usage{
get_names_from_properties(properties)
}
\arguments{
\item{properties}{a claims list from \code{extract_claims}}
}
\value{
tibble of labels for each property for a set of claims
}
\description{
For a claim or set of claims, return the names of the properties
}
