% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{get_random_item}
\alias{get_random_item}
\alias{get_random}
\alias{get_random_property}
\title{Retrieve randomly-selected Wikidata items or properties}
\usage{
get_random_item(limit = 1, ...)

get_random_property(limit = 1, ...)
}
\arguments{
\item{limit}{how many random items to return. 1 by default, but can be higher.}

\item{...}{
  Arguments passed on to \code{\link[httr:GET]{httr::GET}}
  \describe{
    \item{\code{url}}{the url of the page to retrieve}
    \item{\code{config}}{Additional configuration settings such as http
authentication (\code{\link[httr:authenticate]{authenticate()}}), additional headers
(\code{\link[httr:add_headers]{add_headers()}}), cookies (\code{\link[httr:set_cookies]{set_cookies()}}) etc.
See \code{\link[httr:config]{config()}} for full details and list of helpers.}
    \item{\code{handle}}{The handle to use with this request. If not
supplied, will be retrieved and reused from the \code{\link[httr:handle_pool]{handle_pool()}}
based on the scheme, hostname and port of the url. By default \pkg{httr}
requests to the same scheme/host/port combo. This substantially reduces
connection time, and ensures that cookies are maintained over multiple
requests to the same host. See \code{\link[httr:handle_pool]{handle_pool()}} for more
details.}
  }}
}
\description{
\code{get_random_item} and \code{get_random_property} allow you to retrieve the data
associated with randomly-selected Wikidata items and properties, respectively. As with
other \code{WikidataR} code, custom print methods are available; use \code{\link{str}}
to manipulate and see the underlying structure of the data.
}
\examples{
\dontrun{
#Random item
random_item <- get_random_item()

#Random property
random_property <- get_random_property()
}
}
\seealso{
\code{\link{get_item}} for selecting a specific item or property,
or \code{\link{find_item}} for using search functionality to pull out
item or property IDs where the descriptions or aliases match a particular
search term.
}
