% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prints.R
\name{print.find_property}
\alias{print.find_property}
\title{Print method for find_property}
\usage{
\method{print}{find_property}(x, ...)
}
\arguments{
\item{x}{find_property object with search results}

\item{\dots}{Arguments to be passed to methods}
}
\description{
Print method for find_property
}

