% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prints.R
\name{print.wditem}
\alias{print.wditem}
\title{Print method for Wikidata items}
\usage{
\method{print}{wditem}(x, ...)
}
\arguments{
\item{x}{item object from get_item or get_random_item}

\item{\dots}{Arguments to be passed to methods}
}
\description{
Print method for Wikidata items
}

