% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prints.R
\name{print.wikidata}
\alias{print.wikidata}
\title{Print method for Wikidata objects}
\usage{
\method{print}{wikidata}(x, ...)
}
\arguments{
\item{x}{wikidata object from get_item, get_random_item, get_property or get_random_property}

\item{\dots}{Arguments to be passed to methods}
}
\description{
Print method for Wikidata objects
}
\seealso{
get_item, get_random_item, get_property or get_random_property
}

