% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schol.R
\name{qid_from_DOI}
\alias{qid_from_DOI}
\title{QID from DOI}
\usage{
qid_from_DOI(DOI = "10.15347/WJM/2019.001")
}
\arguments{
\item{DOI}{digital object identifiers submitted as strings}
}
\value{
vector of QIDs corresponding to DOIs submitted
}
\description{
simple converter from DOIs to QIDs (for items in wikidata)
}
