% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{qid_from_identifier}
\alias{qid_from_identifier}
\title{QID from identifier}
\usage{
qid_from_identifier(
  property = "DOI",
  value = c("10.15347/WJM/2019.001", "10.15347/WJM/2020.002")
)
}
\arguments{
\item{property}{The identifier property to search (for caveats, see \code{as_pid}.)}

\item{value}{The identifier value to match.}
}
\value{
A vector of QIDs corresponding to identifiers submitted.
}
\description{
Convert unique identifiers to QIDs (for items in Wikidata).
}
\details{
The \href{https://www.wikidata.org/wiki/Q43649390}{Wikidata Q identifier} (QID) is the unique identifier (UID)
used in Wikidata.
}
\examples{
qid_from_identifier('ISBN-13','978-0-262-53817-6')
}
