% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schol.R
\name{qid_from_name}
\alias{qid_from_name}
\title{QID from label name}
\usage{
qid_from_name(name = "Thomas Shafee", limit = 100, format = "vector")
}
\arguments{
\item{name}{name labels submitted as strings}

\item{limit}{if multiple QIDs match each submitted name, how many to return}

\item{format}{output format ('vector' to return a simple vector, or 'list' to return a nested list)}
}
\value{
vector of QIDs corresponding to names submitted. Note: some names may return multiple QIDs.
}
\description{
simple converter from label names to QIDs (for items in wikidata).
Essentially a simplification of \code{find_item}
}
