% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schol.R
\name{qid_from_ORCID}
\alias{qid_from_ORCID}
\title{QID from ORCID}
\usage{
qid_from_ORCID(ORCID = "0000-0002-2298-7593")
}
\arguments{
\item{ORCID}{digital object identifiers submitted as strings}
}
\value{
vector of QIDs corresponding to ORCIDs submitted
}
\description{
simple converter from ORCIDs to QIDs (for items in wikidata)
}
