% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unspecial}
\alias{unspecial}
\title{Remove special characters}
\usage{
unspecial(x)
}
\arguments{
\item{x}{a vector of strings to check for special characters}
}
\value{
the inputted strings with special characters replaced with
closest match plan characters.
}
\description{
Special characters can otherwise mess up wikidata read-writes
}
