% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_quot}
\alias{as_quot}
\title{Add quotations marks}
\usage{
as_quot(x, format = "tibble")
}
\arguments{
\item{x}{a vector, data frame, or tibble of strings}

\item{format}{either "tibble" / "csv" to use plain quotation marks (default), or "api" / "website" to use '\%22'}
}
\value{
tibble of items inside of escaped quotation marks
unless they are already in escaped quotation marks, is a QID, (in which chase it is returned unchanged)
}
\description{
Add escaped quotation marks around strings that need them ready for submision to an API
}
\examples{
as_quot("text")
}
