% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createrows.tidy}
\alias{createrows.tidy}
\title{"CREATE" rows from tidy format}
\usage{
createrows.tidy(QS.tib)
}
\arguments{
\item{QS.tib}{a tibble of items, values and properties (optionally qualifiers and sources).}
}
\value{
a tibble, with items that start with "CREATE" followed by any unique text causing the
addition of a "Create" line above, being replaced with "LAST" in the Quickstatemnts format
to create new QIDs.
}
\description{
Add in QuickStatements CREATE rows that mint new QIDs from tidy input data.
New items are created by any item starting that starts with the text "CREATE" followed
by any unique ID.
}
