% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{find_item}
\alias{find_item}
\alias{find_property}
\title{Search for Wikidata items or properties that match a search term}
\usage{
find_item(
  search_term,
  language = "en",
  limit = 10,
  response_language = "en",
  ...
)

find_property(
  search_term,
  language = "en",
  response_language = "en",
  limit = 10
)
}
\arguments{
\item{search_term}{A term to search for.}

\item{language}{The language to return the labels and descriptions in; this should
consist of an ISO language code. Defaults to \code{"en"}.}

\item{limit}{The number of results to return; set to \code{10} by default.}

\item{\\dots}{further arguments to pass to de{httr:ink[httr::GET]{GET}}.}
}
\value{
A list containing the result of the query.
}
\description{
\code{find_item} and \code{find_property} allow you to retrieve a set
of Wikidata items or properties where the aliases or descriptions match a particular
search term.  As with other \code{WikidataR} code, custom print methods are available;
use \code{\link{str}} to manipulate and see the underlying structure of the data.
}
\examples{

#Check for entries relating to Douglas Adams in some way
adams_items <- find_item("Douglas Adams")

#Check for properties involving the peerage
peerage_props <- find_property("peerage")

}
\seealso{
\code{\link{get_random}} for selecting a random item or property,
or \code{\link{get_item}} for selecting a specific item or property.
}
