% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{find_item}
\alias{find_item}
\alias{find_property}
\title{Search for Wikidata items or properties that match a search term}
\usage{
find_item(
  search_term,
  language = "en",
  limit = 10,
  response_language = "en",
  ...
)

find_property(
  search_term,
  language = "en",
  response_language = "en",
  limit = 10
)
}
\arguments{
\item{search_term}{A term to search for.}

\item{language}{The language to return the labels and descriptions in; this should
consist of an ISO language code. Defaults to \code{"en"}.}

\item{limit}{The number of results to return; set to \code{10} by default.}

\item{...}{
  Arguments passed on to \code{\link[httr:GET]{httr::GET}}
  \describe{
    \item{\code{url}}{the url of the page to retrieve}
    \item{\code{config}}{Additional configuration settings such as http
authentication (\code{\link[httr:authenticate]{authenticate()}}), additional headers
(\code{\link[httr:add_headers]{add_headers()}}), cookies (\code{\link[httr:set_cookies]{set_cookies()}}) etc.
See \code{\link[httr:config]{config()}} for full details and list of helpers.}
    \item{\code{handle}}{The handle to use with this request. If not
supplied, will be retrieved and reused from the \code{\link[httr:handle_pool]{handle_pool()}}
based on the scheme, hostname and port of the url. By default \pkg{httr}
requests to the same scheme/host/port combo. This substantially reduces
connection time, and ensures that cookies are maintained over multiple
requests to the same host. See \code{\link[httr:handle_pool]{handle_pool()}} for more
details.}
  }}
}
\value{
A list containing the result of the query.
}
\description{
\code{find_item} and \code{find_property} allow you to retrieve a set
of Wikidata items or properties where the aliases or descriptions match a particular
search term.  As with other \code{WikidataR} code, custom print methods are available;
use \code{\link{str}} to manipulate and see the underlying structure of the data.
}
\examples{

#Check for entries relating to Douglas Adams in some way
adams_items <- find_item("Douglas Adams")

#Check for properties involving the peerage
peerage_props <- find_property("peerage")

}
\seealso{
\code{\link{get_random}} for selecting a random item or property,
or \code{\link{get_item}} for selecting a specific item or property.
}
