% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{get_geo_box}
\alias{get_geo_box}
\title{Get geographic entities based on a bounding box}
\usage{
get_geo_box(
  first_city_code,
  first_corner,
  second_city_code,
  second_corner,
  language = "en",
  ...
)
}
\arguments{
\item{first_city_code}{a Wikidata item, or series of items, to use for
one corner of the bounding box.}

\item{first_corner}{the direction of \code{first_city_code} relative
to \code{city} (eg "NorthWest", "SouthEast").}

\item{second_city_code}{a Wikidata item, or series of items, to use for
one corner of the bounding box.}

\item{second_corner}{the direction of \code{second_city_code} relative
to \code{city} (eg "NorthWest", "SouthEast").}

\item{language}{the two-letter language code to use for the name
of the item. "en" by default.}

\item{...}{
  Arguments passed on to \code{\link[httr:GET]{httr::GET}}
  \describe{
    \item{\code{url}}{the url of the page to retrieve}
    \item{\code{config}}{Additional configuration settings such as http
authentication (\code{\link[httr:authenticate]{authenticate()}}), additional headers
(\code{\link[httr:add_headers]{add_headers()}}), cookies (\code{\link[httr:set_cookies]{set_cookies()}}) etc.
See \code{\link[httr:config]{config()}} for full details and list of helpers.}
    \item{\code{handle}}{The handle to use with this request. If not
supplied, will be retrieved and reused from the \code{\link[httr:handle_pool]{handle_pool()}}
based on the scheme, hostname and port of the url. By default \pkg{httr}
requests to the same scheme/host/port combo. This substantially reduces
connection time, and ensures that cookies are maintained over multiple
requests to the same host. See \code{\link[httr:handle_pool]{handle_pool()}} for more
details.}
  }}
}
\value{
a data.frame of 5 columns:
\describe{
\item{item}{ the Wikidata identifier of each object associated with
\code{entity}.}
\item{name}{ the name of the item, if available, in the requested language. If it
is not available, \code{NA} will be returned instead.}
\item{latitude}{ the latitude of \code{item}}
\item{longitude}{ the longitude of \code{item}}
\item{entity}{ the entity the item is associated with (necessary for multi-entity
queries).}
}
}
\description{
\code{get_geo_box} retrieves all geographic entities in
Wikidata that fall between a bounding box between two existing items
with geographic attributes (usually cities).
}
\examples{
# Simple bounding box
\donttest{bruges_box <- get_geo_box("Q12988", "NorthEast", "Q184287", "SouthWest")}

# Custom language
\donttest{bruges_box_fr <- get_geo_box("Q12988", "NorthEast", "Q184287", "SouthWest",
                                        language = "fr")}

}
\seealso{
\code{\link{get_geo_entity}} for using an unrestricted search or simple radius,
rather than a bounding box.
}
