% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{get_geo_entity}
\alias{get_geo_entity}
\title{Retrieve geographic information from Wikidata}
\usage{
get_geo_entity(entity, radius = NULL, ...)
}
\description{
\code{get_geo_entity} retrieves any object
with geographic data associated with a geographically-tagged entity
(for example: a city).
}

