% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gets.R
\name{get_random_item}
\alias{get_random}
\alias{get_random_item}
\alias{get_random_property}
\title{Retrieve randomly-selected Wikidata items or properties}
\usage{
get_random_item(...)

get_random_property(...)
}
\arguments{
\item{...}{arguments to pass to httr's GET.}
}
\description{
\code{get_random_item} and \code{get_random_property} allow you to retrieve the data
associated with randomly-selected Wikidata items and properties, respectively. As with
other \code{WikidataR} code, custom print methods are available; use \code{\link{str}}
to manipulate and see the underlying structure of the data.
}
\examples{
#Random item
random_item <- get_random_item()

#Random property
random_property <- get_random_property()
}
\seealso{
\code{\link{get_item}} for selecting a specific item or property,
or \code{\link{find_item}} for using search functionality to pull out
item or property IDs where the descriptions or aliases match a particular
search term.
}

