% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{print.find_item}
\alias{print.find_item}
\title{Print method for find_item}
\usage{
\method{print}{find_item}(x, ...)
}
\arguments{
\item{x}{find_item object with search results}

\item{\dots}{Arguments to be passed to methods}
}
\description{
print found items.
}

