% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{qid_from_identifier}
\alias{qid_from_identifier}
\title{QID from identifier}
\usage{
qid_from_identifier(
  property = "DOI",
  value = c("10.15347/WJM/2019.001", "10.15347/WJM/2020.002")
)
}
\arguments{
\item{property}{the identifier property to search (for caveats, see \code{as_pid})}

\item{value}{the identifier value to match}
}
\value{
vector of QIDs corresponding to identifiers submitted
}
\description{
convert unique identifiers to QIDs (for items in wikidata).
}
\examples{
qid_from_identifier('ISBN-13','978-0-262-53817-6')
}
