% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{sparql_query}
\alias{sparql_query}
\title{Download full Wikidata items matching a sparql query}
\usage{
sparql_query(query, ...)
}
\arguments{
\item{query}{the sparql query as a string}

\item{\\dots}{Additional parameters to supply to [httr::POST]}
}
\value{
a download of the full wikidata objects formatted as a nested json list
}
\description{
Utility wrapper for wikidata spargle endpoint to download items.
Used by \code{get_geo_entity} and \code{get_geo_box}
}
