% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{WD.globalvar}
\alias{WD.globalvar}
\title{Global variables for Wikidata properties}
\format{
A list of tibbles documenting key property constraints from wikidata  
\describe{
  \item{SID.valid}{valid reference source properties}
  \item{PID.datatype}{required data type for each property}
  \item{PID.constraint}{expected regex match for each property}
  \item{lang.abbrev}{language abbreviations}
  \item{lang.abbrev.wiki}{language abbreviations for current wikis}
  \item{abbrev.wiki}{Wikimedia abbreviations for current wikis}
  ...
}
}
\description{
A dataset of Wikidata global variables.
}
