% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{wd_query}
\alias{wd_query}
\title{Download a Wikidata item}
\usage{
wd_query(title, ...)
}
\arguments{
\item{title}{The Wikidata item or property as a string.}

\item{\\dots}{Additional parameters to supply to  \code{httr:\link[httr::POST]{POST}}.}
}
\value{
A downloaded full wikidata object (item or property) formatted as a 
nested json list.
}
\description{
Utility wrapper for Wikidata API to download item.
Used by \code{get_item} and \code{get_property}.
}
