% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{wd_query}
\alias{wd_query}
\title{Download a Wikidata item}
\usage{
wd_query(title, ...)
}
\arguments{
\item{title}{the wikidata item or property as a string}

\item{\\dots}{Additional parameters to supply to [httr::POST]}
}
\value{
a download of the full wikidata object (item or property) formatted as a nested json list
}
\description{
Utility wrapper for wikidata API to download item.
Used by \code{get_item} and \code{get_property}
}
