% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WikidataR.R
\docType{package}
\name{WikidataR}
\alias{WikidataR}
\alias{WikidataR-package}
\title{API client library for Wikidata}
\description{
API client library for Wikidata
}
\seealso{
\code{\link{get_random}} for selecting a random item or property,
\code{\link{get_item}} for a /specific/ item or property, or \code{\link{find_item}}
for using search functionality to pull out item or property IDs where the descriptions
or aliases match a particular search term.
}

